receipt-scanner-ag2/pyproject.toml
[project]
name = "receipt-scanner-ag2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag2[gemini]>=0.9.9",
    "python-dotenv>=1.1.1",
]
